F function code moved to compression package
This will let us to import only F code from projects that require only F
compression function and use govendor to manage dependencies (hello,
go-ethereum!)
