+++ hash.go
@@ -20,6 +20,8 @@ import (
 	"encoding/binary"
 	"errors"
 	"hash"
+
+	"github.com/keep-network/blake2b/compression"
 )
 
 const (
@@ -36,20 +38,9 @@ const (
 )
 
 var (
-	useAVX2 bool
-	useAVX  bool
-	useSSE4 bool
-)
-
-var (
 	errKeySize  = errors.New("blake2b: invalid key size")
 	errHashSize = errors.New("blake2b: invalid hash size")
 )
-
-var iv = [8]uint64{
-	0x6a09e667f3bcc908, 0xbb67ae8584caa73b, 0x3c6ef372fe94f82b, 0xa54ff53a5f1d36f1,
-	0x510e527fade682d1, 0x9b05688c2b3e6c1f, 0x1f83d9abfb41bd6b, 0x5be0cd19137e2179,
-}
 
 // Sum512 returns the BLAKE2b-512 checksum of the data.
 func Sum512(data []byte) [Size]byte {
@@ -115,7 +106,7 @@ func newDigest(hashSize int, key []byte) (*digest, error) {
 }
 
 func checkSum(sum *[Size]byte, hashSize int, data []byte) {
-	h := iv
+	h := compression.IV
 	h[0] ^= uint64(hashSize) | (1 << 16) | (1 << 24)
 	var c [2]uint64
 
@@ -203,7 +194,7 @@ func (d *digest) BlockSize() int { return BlockSize }
 func (d *digest) Size() int { return d.size }
 
 func (d *digest) Reset() {
-	d.h = iv
+	d.h = compression.IV
 	d.h[0] ^= uint64(d.size) | (uint64(d.keyLen) << 8) | (1 << 16) | (1 << 24)
 	d.offset, d.c[0], d.c[1] = 0, 0, 0
 	if d.keyLen > 0 {
@@ -283,7 +274,7 @@ func hashBlocks(h *[8]uint64, blocks []byte, c *[2]uint64, f bool) {
 			i += 8
 		}
 
-		F(h, m, [2]uint64{c0, c1}, f, Rounds)
+		compression.F(h, m, [2]uint64{c0, c1}, f, Rounds)
 	}
 	c[0], c[1] = c0, c1
 }
